+++ src/lib/components/Form/FieldsList.jsx
@@ -1,11 +1,13 @@
-import { Fragment } from 'react'
+import { Fragment, useMemo } from 'react'
 import PropTypes from 'prop-types'
 import { Form as AntdForm, Button, Row, Divider } from 'antd'
-import { MinusCircleOutlined, PlusOutlined } from '@ant-design/icons'
-
-import Field from './Field'
-
+import { PlusOutlined } from '@ant-design/icons'
 import string from '~su/utilities/string'
+
+import getContainer from './utilities'
+import Field from './Field'
+import { DeleteIcon } from './index.styled'
+import Space from '../Space'
 
 // type Props = {
 //   fields: any
@@ -26,33 +28,39 @@ const FieldsList = ({
   separateItems = false,
   rules = [],
   className,
-  onAdd
+  onAdd,
+  container,
+  boldLabels,
+  horizontal = false
 }) => {
   const renderFields = (_formFields, _controlActions, extras = {}) => {
     const fields = dynamic && typeof fieldsConfig === 'function' ? fieldsConfig(extras) : fieldsConfig
 
-    return fields.map((field, index) => {
-      return (
-        <Field
-          key={index}
-          field={field}
-          index={index}
-          disable={disable || field.item?.disable}
-          dynamic={dynamic}
-          extras={extras}
-        />
-      )
-    })
+    const Layout = horizontal ? Space : Fragment
+
+    return (
+      <Layout>
+        {fields.map((field, index) => (
+          <Field
+            key={index}
+            field={field}
+            index={index}
+            disable={disable || field.item?.disable}
+            dynamic={dynamic}
+            extras={extras}
+            boldLabel={boldLabels}
+          />
+        ))}
+      </Layout>
+    )
   }
 
   const renderDynamicFields = (formFields, { add, remove }, { errors }) => {
     let f = formFields.map((formField, index) => {
       return (
-        <Row gutter={[10, 5]} key={index} style={{ position: 'relative', paddingRight: 20 }}>
+        <Row gutter={[10, 5]} key={index} style={{ position: 'relative', paddingRight: 28 }}>
           {renderFields(formFields, { add, remove }, { formField })}
-          {actions.isRemoveAllowed && (
-            <MinusCircleOutlined onClick={() => remove(formField.name)} style={{ position: 'absolute', right: 5 }} />
-          )}
+          {actions.isRemoveAllowed && <DeleteIcon onClick={() => remove(formField.name)} />}
           {separateItems ? <Divider /> : null}
         </Row>
       )
@@ -79,12 +87,18 @@ const FieldsList = ({
     )
   }
 
+  const Container = useMemo(() => getContainer(container?.type), [container])
   const Wrapper = dynamic ? Fragment : Row
   const wrapperProps = dynamic ? {} : { className, gutter: [10, 5] }
 
   return (
-    <>
-      {showTitle ? <h3>{string.humanize(name, { capitalize: true })}</h3> : null}
+    <Container {...container?.props}>
+      {showTitle && (
+        <>
+          <h3>{string.humanize(name, { capitalize: true })}</h3>
+          {container?.type === 'Card' && <Divider style={{ marginTop: '12px' }} />}
+        </>
+      )}
       <Wrapper {...wrapperProps}>
         {name ? (
           <AntdForm.List name={name} rules={rules}>
@@ -94,7 +108,7 @@ const FieldsList = ({
           renderFields(null, {})
         )}
       </Wrapper>
-    </>
+    </Container>
   )
 }
 
@@ -119,7 +133,13 @@ FieldsList.propTypes = {
   separateItems: PropTypes.bool,
   rules: PropTypes.array,
   className: PropTypes.string,
-  onAdd: PropTypes.func
+  onAdd: PropTypes.func,
+  fieldsContainer: PropTypes.shape({
+    type: PropTypes.oneOf(['Card']),
+    props: PropTypes.object
+  }),
+  boldLabels: PropTypes.bool,
+  horizontal: PropTypes.bool
 }
 
 export default FieldsList