+++ src/lib/components/index.js
@@ -30,6 +30,7 @@ import Badge from './Badge'
 import ButtonGroup from './ButtonGroup'
 import ButtonMenu from './ButtonMenu'
 import CodeBlock from './CodeBlock'
+import DatePicker from './DatePicker'
 import Description from './Description'
 import Divider from './Divider'
 import Dropdown from './Dropdown'
@@ -56,6 +57,7 @@ import ProDescriptions from './ProDescriptions'
 import RequestTable from './RequestTable'
 import ResultsTable from './ResultsTable'
 import Score from './Score'
+import SearchBar from './SearchBar'
 import Space from './Space'
 import Spacer from './Spacer'
 import StateBadge from './StateBadge'
@@ -69,6 +71,7 @@ import Tabs from './Tabs'
 import Tag from './Tag'
 import TagsList from './TagsList'
 import TextWithCountBadge from './TextWithCountBadge'
+import TimePicker from './TimePicker'
 import Timeline from './Timeline'
 import TimeTaken from './TimeTaken'
 import ToggledSelect from './ToggledSelect'
@@ -77,6 +80,7 @@ import Typography from './Typography'
 import YesNoTag from './YesNoTag'
 import CompanyLogo from './CompanyLogo'
 import SmartDataDisplay from './SmartDataDisplay'
+import Radio from './Radio'
 import RadioGroup from './RadioGroup'
 import Breadcrumb from './Breadcrumb'
 import TreeSelect from './TreeSelect'
@@ -114,6 +118,7 @@ export {
   ButtonGroup,
   ButtonMenu,
   CodeBlock,
+  DatePicker,
   Description,
   Divider,
   Dropdown,
@@ -142,6 +147,7 @@ export {
   RequestTable,
   ResultsTable,
   Skeleton,
+  SearchBar,
   Space,
   Spacer,
   StateBadge,
@@ -155,6 +161,7 @@ export {
   Tag,
   TagsList,
   TextWithCountBadge,
+  TimePicker,
   Timeline,
   TimeTaken,
   ToggledSelect,
@@ -163,6 +170,7 @@ export {
   YesNoTag,
   CompanyLogo,
   SmartDataDisplay,
+  Radio,
   RadioGroup,
   Breadcrumb,
   TreeSelect,